+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opModes/autos/blue/BeaconBlue.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.opModes.autos;
+package org.firstinspires.ftc.teamcode.opModes.autos.blue;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
@@ -9,8 +9,8 @@ import org.firstinspires.ftc.teamcode.subsystems.Catapult;
 import org.firstinspires.ftc.teamcode.subsystems.Drivetrain;
 import org.firstinspires.ftc.teamcode.subsystems.Intake;
 
-@Autonomous(name="legend27: SDB Blue", group="legend27")
-public class ShootDriveBeaconBlue extends LinearOpMode {
+@Autonomous(name="legend27: Beacon Blue", group="legend27")
+public class BeaconBlue extends LinearOpMode {
 
     private boolean teamRed = false;
 
@@ -25,49 +25,12 @@ public class ShootDriveBeaconBlue extends LinearOpMode {
     public void runOpMode() throws InterruptedException {
         timer = new ElapsedTime();
 
-        catapult = new Catapult(hardwareMap.dcMotor.get("catapult"));
         drivetrain = new Drivetrain(hardwareMap.dcMotor.get("left_drive"), hardwareMap.dcMotor.get("right_drive"));
         intake = new Intake(hardwareMap.dcMotor.get("intake"));
+        catapult = new Catapult(hardwareMap.dcMotor.get("catapult"));
         beaconActivator = new BeaconActivator(hardwareMap.colorSensor.get("sensor"), hardwareMap.deviceInterfaceModule.get("dim"), hardwareMap.led.get("led"), hardwareMap.dcMotor.get("button_pusher"));
 
         waitForStart();
-
-        drivetrain.tankDrive(1, 1);
-        telemetry.addData("Status", "Driving forward0");
-        timer.reset();
-        while (timer.seconds() < .1 && opModeIsActive());
-        drivetrain.stop();
-
-        if(!teamRed) {
-            drivetrain.tankDrive(-0.125, 0.125);
-        }
-        else {
-            drivetrain.tankDrive(0.125, -0.125);
-        }
-
-        telemetry.addData("Status", "Turning");
-
-        timer.reset();
-        while (timer.seconds() < 0.57 && opModeIsActive());
-        drivetrain.stop();
-
-        catapult.catapultBall(1.05, .25);
-
-        timer.reset();
-        while(timer.seconds() < 2);
-
-        catapult.catapultBall(1.05, .25);
-
-        if(teamRed) {
-            drivetrain.tankDrive(-0.125, 0.125);
-        }
-        else {
-            drivetrain.tankDrive(0.125, -0.125);
-        }
-        telemetry.addData("Status", "Turning");
-
-        timer.reset();
-        while (timer.seconds() < 0.47 && opModeIsActive());
 
         drivetrain.tankDrive(1, 1);
         telemetry.addData("Status", "Driving forward0");
@@ -83,12 +46,12 @@ public class ShootDriveBeaconBlue extends LinearOpMode {
         telemetry.addData("Status", "Turning");
 
         timer.reset();
-        while (timer.seconds() < 2.35 && opModeIsActive());
+        while (timer.seconds() < 2.45 && opModeIsActive());
 
         drivetrain.tankDrive(.2, .2);
         telemetry.addData("Status", "Driving forward1 ");
         timer.reset();
-        while (timer.seconds() < 3.5 && opModeIsActive());
+        while (timer.seconds() < 3 && opModeIsActive());
 
         drivetrain.stop();
